+++ server/src/common/mod.rs
@@ -1,0 +1,5 @@
+pub mod errors;
+
+pub trait Validation {
+  fn validate(&self) -> Result<(), String>;
+}
