--- server/src/services.rs
@@ -1,2 +1,0 @@
-pub mod db_service;
-pub mod plaid;
