+++ server/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod db;
+pub mod plaid;
